+++ src/main/java/com/samsamhajo/deepground/entity/question/Question.java
@@ -1,8 +1,7 @@
-package com.samsamhajo.deepground.entity;
+package com.samsamhajo.deepground.entity.question;
 
 import com.samsamhajo.deepground.global.BaseEntity;
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
